+++ src/options/PuckSettings.tsx
@@ -3,7 +3,11 @@ import { useCallback } from 'preact/hooks';
 import type { Config } from '../common/config';
 import { useLocale } from '../common/i18n';
 
-import { PuckSettingsForm, type ShowPuckSetting } from './PuckSettingsForm';
+import {
+  type HandednessSetting,
+  PuckSettingsForm,
+  type ShowPuckSetting,
+} from './PuckSettingsForm';
 import { SectionHeading } from './SectionHeading';
 import { useConfigValue } from './use-config-value';
 
@@ -12,9 +16,16 @@ type Props = { config: Config };
 export function PuckSettings(props: Props) {
   const { t } = useLocale();
   const showPuck = useConfigValue(props.config, 'showPuck');
+  const handedness = useConfigValue(props.config, 'handedness');
   const onChangeShowPuck = useCallback(
     (value: ShowPuckSetting) => {
       props.config.showPuck = value;
+    },
+    [props.config]
+  );
+  const onChangeHandedness = useCallback(
+    (value: HandednessSetting) => {
+      props.config.handedness = value;
     },
     [props.config]
   );
@@ -23,7 +34,12 @@ export function PuckSettings(props: Props) {
     <>
       <SectionHeading>{t('options_lookup_puck_heading')}</SectionHeading>
       <div class="py-4">
-        <PuckSettingsForm showPuck={showPuck} onChange={onChangeShowPuck} />
+        <PuckSettingsForm
+          showPuck={showPuck}
+          onChangeShowPuck={onChangeShowPuck}
+          handedness={handedness}
+          onChangeHandedness={onChangeHandedness}
+        />
       </div>
     </>
   );