+++ CHANGELOG.md
@@ -11,6 +11,9 @@ app.
 - Added references for
   [The Kodansha Kanji Dictionary](https://www.kanji.org/dictionaries/KKD/kaneirev.htm).
 - Improved scanning of ruby transcription text.
+- Added vertical reading mode to puck, allowing puck earth to orbit on the sides
+  rather than just the top or bottom
+  ([#2557](https://github.com/birchill/10ten-ja-reader/pull/2557)).
 - Revert changes from [PR #1773](https://github.com/birchill/10ten-ja-reader/pull/1773).
   Puck can hold popup open again ([#2564](https://github.com/birchill/10ten-ja-reader/pull/2564)).
 - Made popup work for Notion popups like the calendar